XD-2499: Add partitionResultsTimeout to metadata
